pyproject.toml
[project]
authors = [
    {name = "TobyBoyne", email = "tobyboyne@gmail.com"},
]
license = {text = "LICENSE"}
requires-python = "<4.0,>=3.10"
dependencies = [
    "torch<3.0.0,>=2.2.0",
    "numpy<2.0.0,>=1.24.4",
    "matplotlib<4.0.0,>=3.8.2",
    "gpytorch<2.0,>=1.11",
    "tqdm<5.0.0,>=4.66.1",
    "scikit-optimize<1.0.0,>=0.9.0",
    "gurobipy==12",
    "lightgbm<5.0.0,>=4.3.0",
    "botorch==0.10.0",
    "beartype<1.0.0,>=0.17.2",
    "ucimlrepo<1.0.0,>=0.0.6",
    "numba<1.0.0,>=0.60.0",
    "xgboost<3.0.0,>=2.1.1",
    "bofire[cheminfo,optimization]<1.0.0,>=0.0.14",
    "numba-scipy<1.0.0,>=0.4.0",
    "entmoot<3.0.0,>=2.0.2",
    "arviz>=0.21.0",
    "pymc-bart>=0.9.2",
]
name = "bark"
version = "0.1.0"
description = ""
readme = "README.md"


[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I001"]
ignore = ["E721", "E731", "F722", "F821"]
ignore-init-module-imports = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Defines the targets for the wheel build.
include = ["src/bark/**", "src/bofire_mixed/**"]
sources = ["src"]
packages = ["src/bark", "src/bofire_mixed"]


[dependency-groups]
dev = [
    "pytest<9.0.0,>=8.0.0",
    "ruff<1.0.0,>=0.2.0",
    "pre-commit<4.0.0,>=3.6.0",
    "ipython<9.0.0,>=8.21.0",
    "pytest-benchmark<5.0.0,>=4.0.0",
]
